Tests updates; Document with Global Roles - state
