+++ test/non-functional/auth/my-privileges/global-roles/global-admin/hub.it-spec.ts
@@ -143,7 +143,7 @@ describe('myPrivileges', () => {
           .myPrivileges;
 
       // Assert
-      expect(data).toEqual(['CREATE', 'GRANT', 'READ', 'UPDATE', 'DELETE']);
+      expect(data).toEqual(cgrud);
     });
 
     test('GlobalAdmin privileges to Hub / Community / Communication', async () => {
@@ -154,7 +154,7 @@ describe('myPrivileges', () => {
           .myPrivileges;
 
       // Assert
-      expect(data).toEqual(['CREATE', 'GRANT', 'READ', 'UPDATE', 'DELETE']);
+      expect(data).toEqual(cgrud);
     });
 
     test('GlobalAdmin privileges to Hub / Community / Communication / Discussion', async () => {
@@ -165,7 +165,7 @@ describe('myPrivileges', () => {
           .authorization.myPrivileges;
 
       // Assert
-      expect(data).toEqual(['CREATE', 'GRANT', 'READ', 'UPDATE', 'DELETE']);
+      expect(data).toEqual(cgrud);
     });
 
     test('GlobalAdmin privileges to Hub / Community / Communication / Updates', async () => {
@@ -176,7 +176,7 @@ describe('myPrivileges', () => {
           .myPrivileges;
 
       // Assert
-      expect(data).toEqual(['CREATE', 'GRANT', 'READ', 'UPDATE', 'DELETE']);
+      expect(data).toEqual(cgrud);
     });
   });
 
@@ -316,7 +316,7 @@ describe('myPrivileges', () => {
       const data = response.body.data.hub.templates.authorization.myPrivileges;
 
       // Assert
-      expect(data).toEqual(['CREATE', 'GRANT', 'READ', 'UPDATE', 'DELETE']);
+      expect(data).toEqual(cgrud);
     });
 
     test('GlobalAdmin privileges to Hub / Templates / Aspect', async () => {
@@ -327,7 +327,7 @@ describe('myPrivileges', () => {
           .myPrivileges;
 
       // Assert
-      expect(data).toEqual(['CREATE', 'GRANT', 'READ', 'UPDATE', 'DELETE']);
+      expect(data).toEqual(cgrud);
     });
 
     test('GlobalAdmin privileges to Hub / Templates / Lifecycle', async () => {
@@ -338,11 +338,11 @@ describe('myPrivileges', () => {
           .myPrivileges;
 
       // Assert
-      expect(data).toEqual(['CREATE', 'GRANT', 'READ', 'UPDATE', 'DELETE']);
+      expect(data).toEqual(cgrud);
     });
 
     // ToDo
-    test.skip('GlobalAdmin privileges to Hub / Templates / Lifecycle', async () => {
+    test.skip('GlobalAdmin privileges to Hub / Templates / Canvas', async () => {
       // Act
       const response = await getHubData(entitiesId.hubId);
       const data =
